requirements.txt
azure-storage-blob=12.8.0
azure-storage=0.36.0